fix(lab4): qemu launch options
if both `-serial stdio` and `-display curses` are specified to start
qemu, the first n keyboard ints will not be triggered, where n refers
to the size of the COM1 FIFO, which is 14 in the current env.

remove `-serial stdio` or use other display mode would deal the problem,
but may be not a perfect solution.